21/09

+++ Lista2/Ex13Lista2.java
@@ -7,26 +7,26 @@ public class Ex13Lista2 {
         Scanner input = new Scanner(System.in);
 
         System.out.print("Digite o primeiro número: ");
-        double n1 = input.nextDouble();
+        float n1 = input.nextFloat();
 
         System.out.print("Digite o sinal da operação (+, -, *, /): ");
         char operacao = input.next().charAt(0);
 
         System.out.print("Digite o segundo número: ");
-        double n2 = input.nextDouble();
+        float n2 = input.nextFloat();
 
         if (operacao == '+') {
-            double resultado = n1 + n2;
+            float resultado = n1 + n2;
             System.out.printf("%.2f + %.2f = %.2f", n1, n2, resultado);
         } else if (operacao == '-') {
-            double resultado = n1 - n2;
+            float resultado = n1 - n2;
             System.out.printf("%.2f + %.2f = %.2f", n1, n2, resultado);
         } else if (operacao == '*') {
-            double resultado = n1 * n2;
+            float resultado = n1 * n2;
             System.out.printf("%.2f + %.2f = %.2f", n1, n2, resultado);
         } else if (operacao == '/') {
             if (n2 > 0) {
-                double resultado = n1 / n2;
+                float resultado = n1 / n2;
                 System.out.printf("%.2f / %.2f = %.2f", n1, n2, resultado);
             } else {
                 System.out.println("Impossível dividir!");
